requirements.txt
python-dotenv==1.0.1
terminaltables==3.1.10
requests==2.32.3